Moved upgrades construction to Game
